Cleaned up code formatting, plus some other stuff
Hey, I cleaned up the general formatting of the code (mix of tabs and spaces for indents due to copy pasting, so I made it all spaces), changed the Add Contact page so it shouldn't erase the info in the boxes if the contact is invalid (the user may have just made a mistake, and might find it annoying to have to retype it), slightly changed come of the text on the webpage, and added a rounded border to the result divs.

I tried to change the error text parts from red text to a red border, but the border was showing up even when there wasn't any text in it, so I wasn't sure how to make that work.